Make KakapoServer internal, always use Router for registering, unregistering and disabling
